image.bbclass: do not create image manifest link if IMAGE_LINK_NAME is empty

+++ meta/classes/image.bbclass
@@ -322,8 +322,10 @@ make_zimage_symlink_relative () {
 write_image_manifest () {
 	rootfs_${IMAGE_PKGTYPE}_write_manifest
 
-	rm -f ${DEPLOY_DIR_IMAGE}/${IMAGE_LINK_NAME}.manifest
-        ln -s ${IMAGE_NAME}.rootfs.manifest ${DEPLOY_DIR_IMAGE}/${IMAGE_LINK_NAME}.manifest
+	if [ -n "${IMAGE_LINK_NAME}" ]; then
+		rm -f ${DEPLOY_DIR_IMAGE}/${IMAGE_LINK_NAME}.manifest
+		ln -s ${IMAGE_NAME}.rootfs.manifest ${DEPLOY_DIR_IMAGE}/${IMAGE_LINK_NAME}.manifest
+	fi
 }
 
 # Make login manager(s) enable automatic login.
